+++ src/services/permissions/checks/domains/person/xRayBodyScans/XRayBodyScansCheck.ts
@@ -1,16 +1,15 @@
 import PrisonerPermissionsContext from '../../../../../../types/internal/permissions/PrisonerPermissionsContext'
-import xrbsReadCheck from './xrbsRead/XRBSReadCheck'
 import {
   XRayBodyScansPermission,
   XRayBodyScansPermissions,
 } from '../../../../../../types/public/permissions/domains/person/xRayBodyScans/XRayBodyScansPermissions'
-import xrbsEditCheck from './xrbsEdit/XRBSEditCheck'
 import { checkWith } from '../../../../utils/PermissionCheckUtils'
+import xrbsReadAndEditCheck from './XRBSReadAndEditCheck'
 
 export default function xRayBodyScansCheck(context: PrisonerPermissionsContext): XRayBodyScansPermissions {
   const check = checkWith(context)
   return {
-    ...check(XRayBodyScansPermission.read_scans, xrbsReadCheck),
-    ...check(XRayBodyScansPermission.edit_scans, xrbsEditCheck),
+    ...check(XRayBodyScansPermission.read_scans, xrbsReadAndEditCheck),
+    ...check(XRayBodyScansPermission.edit_scans, xrbsReadAndEditCheck),
   }
 }