+++ src/services/permissions/utils/PermissionUtils.test.ts
@@ -1,4 +1,5 @@
 import {
+  checkTimeBasedAccessPostTransfer,
   isActiveCaseLoad,
   isInUsersCaseLoad,
   isReleased,
@@ -21,7 +22,9 @@ import { PersonSentenceCalculationPermission } from '../../../types/public/permi
 import { Role } from '../../../types/internal/user/Role'
 import Prisoner from '../../../data/hmppsPrisonerSearch/interfaces/Prisoner'
 import { isGranted } from '../../../types/public/permissions/prisoner/PrisonerPermissionsUtils'
+import { fixedClock } from '../../../testUtils/fixedClock'
 import { prisonerPermissionsMock } from '../../../testUtils/PrisonerPermissionsMock'
+import { daysToMilliseconds } from './DateUtils'
 
 describe('PermissionUtils', () => {
   describe('isRequiredPermission', () => {
@@ -50,7 +53,7 @@ describe('PermissionUtils', () => {
     `(
       'roles to check: $roles | user roles: $userRoles | user has some roles: $result',
       async ({ roles, userRoles, result }) => {
-        expect(userHasSomeRolesFrom(roles, { userRoles } as HmppsUser)).toEqual(result)
+        expect(userHasSomeRolesFrom(roles, { userRoles } as HmppsUser)).toBe(result)
       },
     )
   })
@@ -68,7 +71,7 @@ describe('PermissionUtils', () => {
     `(
       'roles to check: $roles | user roles: $userRoles | user has all roles: $result',
       async ({ roles, userRoles, result }) => {
-        expect(userHasAllRoles(roles, { userRoles } as HmppsUser)).toEqual(result)
+        expect(userHasAllRoles(roles, { userRoles } as HmppsUser)).toBe(result)
       },
     )
   })
@@ -83,7 +86,7 @@ describe('PermissionUtils', () => {
     `(
       'roles to check: $roles | user roles: $userRoles | user has role: $result',
       async ({ role, userRoles, result }) => {
-        expect(userHasRole(role, { userRoles } as HmppsUser)).toEqual(result)
+        expect(userHasRole(role, { userRoles } as HmppsUser)).toBe(result)
       },
     )
   })
@@ -91,50 +94,50 @@ describe('PermissionUtils', () => {
   describe('isActiveCaseLoad', () => {
     it('Should return true when the prisonId matches the active case load', () => {
       const user = { authSource: 'nomis', activeCaseLoadId: 'ABC' } as PrisonUser
-      expect(isActiveCaseLoad('ABC', user)).toEqual(true)
+      expect(isActiveCaseLoad('ABC', user)).toBe(true)
     })
 
     it('Should return false when the prisonId does not match the active case load', () => {
       const user = { authSource: 'nomis', activeCaseLoadId: 'ABC' } as PrisonUser
-      expect(isActiveCaseLoad('DEF', user)).toEqual(false)
+      expect(isActiveCaseLoad('DEF', user)).toBe(false)
     })
 
     it('Should return false for non prison users', () => {
       const probationUser = { authSource: 'delius' } as ProbationUser
       const externalUser = { authSource: 'external' } as ExternalUser
 
-      expect(isActiveCaseLoad('123', probationUser)).toEqual(false)
-      expect(isActiveCaseLoad('123', externalUser)).toEqual(false)
+      expect(isActiveCaseLoad('123', probationUser)).toBe(false)
+      expect(isActiveCaseLoad('123', externalUser)).toBe(false)
     })
   })
 
   describe('isInUsersCaseLoad', () => {
     it('Should return true when the user has a caseload matching the prisoner', () => {
       const caseLoads: CaseLoad[] = [
-        { caseloadFunction: '', caseLoadId: 'ABC', currentlyActive: false, description: '', type: '' },
-        { caseloadFunction: '', caseLoadId: 'DEF', currentlyActive: false, description: '', type: '' },
+        { caseloadFunction: 'GENERAL', caseLoadId: 'ABC', currentlyActive: false, description: '', type: 'INST' },
+        { caseloadFunction: 'GENERAL', caseLoadId: 'DEF', currentlyActive: false, description: '', type: 'INST' },
       ]
       const user = { authSource: 'nomis', caseLoads } as PrisonUser
 
-      expect(isInUsersCaseLoad('DEF', user)).toEqual(true)
+      expect(isInUsersCaseLoad('DEF', user)).toBe(true)
     })
 
     it('Should return false when the user has a caseload that doesnt match the prisoner', () => {
       const caseLoads: CaseLoad[] = [
-        { caseloadFunction: '', caseLoadId: 'ABC', currentlyActive: false, description: '', type: '' },
-        { caseloadFunction: '', caseLoadId: 'DEF', currentlyActive: false, description: '', type: '' },
+        { caseloadFunction: 'GENERAL', caseLoadId: 'ABC', currentlyActive: false, description: '', type: 'INST' },
+        { caseloadFunction: 'GENERAL', caseLoadId: 'DEF', currentlyActive: false, description: '', type: 'INST' },
       ]
       const user = { authSource: 'nomis', caseLoads } as PrisonUser
 
-      expect(isInUsersCaseLoad('123', user)).toEqual(false)
+      expect(isInUsersCaseLoad('123', user)).toBe(false)
     })
 
     it('Should return false for non prison users', () => {
       const probationUser = { authSource: 'delius' } as ProbationUser
       const externalUser = { authSource: 'external' } as ExternalUser
 
-      expect(isInUsersCaseLoad('123', probationUser)).toEqual(false)
-      expect(isInUsersCaseLoad('123', externalUser)).toEqual(false)
+      expect(isInUsersCaseLoad('123', probationUser)).toBe(false)
+      expect(isInUsersCaseLoad('123', externalUser)).toBe(false)
     })
   })
 
@@ -144,8 +147,8 @@ describe('PermissionUtils', () => {
       ['TRN', false],
       ['MDI', false],
       ['LEI', false],
-    ])('', (prisonId: string, result: boolean) => {
-      expect(isReleased({ prisonId } as Prisoner)).toEqual(result)
+    ])('for %s should return %s', (prisonId: string, result: boolean) => {
+      expect(isReleased({ prisonId } as Prisoner)).toBe(result)
     })
   })
 
@@ -155,8 +158,8 @@ describe('PermissionUtils', () => {
       ['TRN', true],
       ['MDI', false],
       ['LEI', false],
-    ])('', (prisonId: string, result: boolean) => {
-      expect(isTransferring({ prisonId } as Prisoner)).toEqual(result)
+    ])('for %s should return %s', (prisonId: string, result: boolean) => {
+      expect(isTransferring({ prisonId } as Prisoner)).toBe(result)
     })
   })
 
@@ -166,8 +169,51 @@ describe('PermissionUtils', () => {
       ['TRN', true],
       ['MDI', false],
       ['LEI', false],
-    ])('', (prisonId: string, result: boolean) => {
-      expect(isReleasedOrTransferring({ prisonId } as Prisoner)).toEqual(result)
+    ])('for %s should return %s', (prisonId: string, result: boolean) => {
+      expect(isReleasedOrTransferring({ prisonId } as Prisoner)).toBe(result)
+    })
+  })
+
+  describe('checkTimeBasedAccessPostTransfer', () => {
+    // user in ABC with access to DEF as well
+    const user = {
+      authSource: 'nomis',
+      activeCaseLoadId: 'ABC',
+      caseLoads: [
+        { caseloadFunction: 'GENERAL', caseLoadId: 'ABC', currentlyActive: true, description: '', type: 'INST' },
+        { caseloadFunction: 'GENERAL', caseLoadId: 'DEF', currentlyActive: false, description: '', type: 'INST' },
+      ],
+    } as PrisonUser
+    const gracePeriod = daysToMilliseconds(90)
+
+    beforeAll(() => {
+      fixedClock()
+    })
+
+    it.each([
+      { scenario: 'reject prisoner whose previous prison is not set', prisoner: { prisonId: 'ABC' }, result: false },
+      {
+        scenario: 'reject prisoner whose previous prison is not in caseloads',
+        prisoner: { prisonId: 'GHI', previousPrisonId: 'JKL', previousPrisonLeavingDate: '2026-07-15T10:50' },
+        result: false,
+      },
+      {
+        scenario: 'reject prisoner whose previous prison leaving date is unknown',
+        prisoner: { prisonId: 'GHI', previousPrisonId: 'ABC' },
+        result: false,
+      },
+      {
+        scenario: 'reject prisoner whole previous prison is in case loads but not within grace period',
+        prisoner: { prisonId: 'GHI', previousPrisonId: 'ABC', previousPrisonLeavingDate: '2026-04-24T19:30' },
+        result: false,
+      },
+      {
+        scenario: 'accept prisoner whose previous prison is in case loads and within grace period',
+        prisoner: { prisonId: 'GHI', previousPrisonId: 'DEF', previousPrisonLeavingDate: '2026-07-15T10:50' },
+        result: true,
+      },
+    ])('should $scenario', ({ prisoner, result }) => {
+      expect(checkTimeBasedAccessPostTransfer(user, prisoner as Prisoner, gracePeriod)).toBe(result)
     })
   })